+++ android/sdk/src/main/java/com/taobao/weex/ui/WXRenderManager.java
@@ -204,7 +204,6 @@
  */
 package com.taobao.weex.ui;
 
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.text.TextUtils;
 
@@ -214,7 +213,6 @@ import com.taobao.weex.common.WXThread;
 import com.taobao.weex.dom.RenderAction;
 import com.taobao.weex.dom.RenderActionContext;
 import com.taobao.weex.dom.WXDomObject;
-import com.taobao.weex.ui.animation.WXAnimationBean;
 import com.taobao.weex.ui.component.WXComponent;
 import com.taobao.weex.utils.WXUtils;
 
@@ -323,16 +321,6 @@ public class WXRenderManager {
       return;
     }
     statement.setExtra(ref, extra);
-  }
-
-  public void startAnimation(String instanceId, @NonNull String ref,
-                             @NonNull WXAnimationBean animationBean, @Nullable String
-      callBack) {
-    RenderActionContextImpl statement = mRegistries.get(instanceId);
-    if (statement == null) {
-      return;
-    }
-    statement.startAnimation(ref, animationBean, callBack);
   }
 
   public List<WXSDKInstance> getAllInstances() {