+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 ### Changed
 - [android]Move DOM and render to seperate Action Object
+- [android]Move Animation(animation module and css style) to seperate Action Object
+- [android]Reserve `transformOrigin`'s last time state.